Move generator to pkg/gen
