+++ cmd/cert-generator/main.go
@@ -2,19 +2,13 @@ package main
 
 import (
 	"bufio"
-	"crypto/ecdsa"
-	"crypto/elliptic"
-	"crypto/rand"
 	"crypto/x509"
-	"crypto/x509/pkix"
-	"encoding/pem"
-	"errors"
 	"fmt"
-	"math/big"
 	"os"
 	"strconv"
 	"strings"
-	"time"
+
+	"github.com/lazybark/cert-generator/pkg/gen"
 )
 
 func main() {
@@ -39,6 +33,9 @@ func CLIGenerate() error {
 	fmt.Print("Host ('localhost' if empty) -> ")
 	scanner.Scan()
 	host := scanner.Text()
+	if host == "" {
+		host = "localhost"
+	}
 	// Lifetime
 	fmt.Print("Lifetime (days) -> ")
 	scanner.Scan()
@@ -76,79 +73,23 @@ func CLIGenerate() error {
 	fmt.Print("Certificate path ('cert.pem' if empty) -> ")
 	scanner.Scan()
 	certPath := scanner.Text()
+	if certPath == "" {
+		certPath = "cert"
+	}
+
 	fmt.Print("Key path ('key.pem' if empty) -> ")
 	scanner.Scan()
 	keyPath := scanner.Text()
+	if keyPath == "" {
+		keyPath = "key"
+	}
 
-	err = Generator(keyPath, certPath, orgName, host, lifetime, usageArrayX509)
+	err = gen.Generator(keyPath, certPath, orgName, host, lifetime, usageArrayX509)
 	if err != nil {
 		return err
 	}
 	fmt.Printf("Generated certificate: %s.pem\n", certPath)
 	fmt.Printf("Generated private key: %s.pem\n", keyPath)
-
-	return nil
-}
-
-func Generator(keyPath string, certPath string, orgName string, host string, lifetime int, usageArrayX509 []x509.ExtKeyUsage) error {
-	if keyPath == "" {
-		keyPath = "key"
-	}
-	if certPath == "" {
-		certPath = "cert"
-	}
-	if host == "" {
-		host = "localhost"
-	}
-
-	privateKey, err := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-	if err != nil {
-		return fmt.Errorf("[Generator] failed to generate private key -> %w", err)
-	}
-
-	serialNumberLimit := new(big.Int).Lsh(big.NewInt(1), 128)
-	serialNumber, err := rand.Int(rand.Reader, serialNumberLimit)
-	if err != nil {
-		return fmt.Errorf("[Generator] failed to generate serial number -> %w", err)
-	}
-
-	template := x509.Certificate{
-		SerialNumber: serialNumber,
-		Subject: pkix.Name{
-			Organization: []string{orgName},
-		},
-		DNSNames:              []string{host},
-		NotBefore:             time.Now(),
-		NotAfter:              time.Now().AddDate(0, 0, lifetime),
-		KeyUsage:              x509.KeyUsageDigitalSignature,
-		ExtKeyUsage:           usageArrayX509,
-		BasicConstraintsValid: true,
-	}
-
-	derBytes, err := x509.CreateCertificate(rand.Reader, &template, &template, &privateKey.PublicKey, privateKey)
-	if err != nil {
-		return fmt.Errorf("[Generator] failed to create certificate -> %w", err)
-	}
-
-	pemCert := pem.EncodeToMemory(&pem.Block{Type: "CERTIFICATE", Bytes: derBytes})
-	if pemCert == nil {
-		return errors.New("[Generator] failed to encode certificate to PEM")
-	}
-	if err := os.WriteFile(certPath+".pem", pemCert, 0644); err != nil {
-		return err
-	}
-
-	privBytes, err := x509.MarshalPKCS8PrivateKey(privateKey)
-	if err != nil {
-		return fmt.Errorf("[Generator] unable to marshal private key -> %w", err)
-	}
-	pemKey := pem.EncodeToMemory(&pem.Block{Type: "PRIVATE KEY", Bytes: privBytes})
-	if pemKey == nil {
-		return errors.New("[Generator] failed to encode key to PEM")
-	}
-	if err := os.WriteFile(keyPath+".pem", pemKey, 0600); err != nil {
-		return err
-	}
 
 	return nil
 }
